Remove the old binary search tree folder and move the sweet test method into the new one (temporary!)
